Crawler/requirements.txt
pandas==1.1.4
numpy==1.19.4
urllib3==1.26.2
xlrd==1.2.0
requests==2.25.0
shapely==1.7.1
geopy==2.0.0
geojson==2.5.0
SQLAlchemy==1.3.20
GeoAlchemy2==0.8.4
psycopg2-binary==2.8.6
sentry_sdk==0.19.4
python-dotenv==0.15.0
